Standardize tsconfig files (#458)
This project has two kinds of TypeScript config files, which is
consistent with the module template. However, whereas in the module
template, the default config file (`tsconfig.json`) is the one used for
development and `tsconfig.build.json` is used to build the project, in
this repo it's reversed (`tsconfig.test.json` used for development).

This commit aligns the kinds of TypeScript configuration files with the
module template. It also normalizes the existing options to fit the
module template. There should be no differences to `dist/`.